Adapt A2C algorithm to accept continuous reward
Rather than Boolean reward signals, allow for non-integer rewards. Multiply the log likelihood by a the cumulative reward of a trace. I expect this to be fine, but the original code was very explicit in keep reward integer (but I'm not sure why), so there might be some unexpected consequences...