Speed up intersection-over-union
I calculate intersection-over-union with a 2d convolution, but profiling showed that this was super slow - turned out I was using the *executed* shapes, padded to a fixed resolution, rather than the *rendered* shapes, which are much smaller. Also tried manually shifting objects around instead of convolving, but 2d convolution is much faster.